rename the db -> lib folder [ci skip]
