+++ lib/src/supp_calc.py
@@ -1,6 +1,6 @@
 import logging
 logger = logging.getLogger("pdst-calc") 
-from db.src.dst_calc import *
+from lib.src.dst_calc import *
 from tabulate import tabulate
 # Print and log
 def print_and_log_tabulate(df, *args, **kwargs):
